pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "worldsense"
version = "0.1.0"
description = "AI-powered large-scale user research simulation platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.5",
    "httpx>=0.27",
    "aiofiles>=23.0",
    "tenacity>=8.2",
    "aiosqlite>=0.20",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "pymupdf>=1.27.2.2",
    "python-docx>=1.2.0",
    "openai>=2.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
ws = "worldsense.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["worldsense*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
